update 6 files
